test(worker-sessions): split terminal append observer
